app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "RNB's backend code"
authors = ["RNB team <tech@rnb.beta.gouv.fr>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.13.4"
boto3 = "~1.42.80"
Django = "~6.0.3"
drf-spectacular = "~0.29.0"
drf-api-tracking = "~1.8.4"
flower = "~2.0.1"
sentry-sdk = "~2.57.0"
django-debug-toolbar = "~6.2.0"
djangorestframework-gis = "~1.2"
psycopg2 = "~2.9.11"
django-cors-headers = "~4.9.0"
django-extensions = "~4.1"
django-revproxy = "~0.13.0"
nanoid = "~2.0.0"
geopy = "^2.4.1"
whitenoise = "~6.12.0"
watchdog = {extras = ["watchmedo"], version = "~6.0.0"}
celery = "~5.6.3"
py7zr = "~1.1.0"
ijson = "~3.5.0"
redis = "~7.4.0"
gunicorn = "~25.3.0"
requests = "~2.33.1"
pandas = "~3.0.2"
Fiona = "~1.10.1"
djangorestframework = "~3.16.1"
geopandas = "^1.1.3"
openapi-spec-validator = "^0.8.4"
orjson = "^3.11.8"
django-jsoneditor = "^0.2.4"
django-taggit = "^6.1.0"
private-captcha = "^0.0.6"
rapidfuzz = "^3.14.3"
authlib = "~1.5"

[tool.poetry.group.dev.dependencies]
moto = "^5.1.22"
freezegun = "^1.5.5"
notebook = "==6.5.7"
jupyter = "^1.0.0"
openpyxl = "^3.1.5"
tqdm = "^4.67.3"
mypy = "==1.19.1"
django-stubs = {extras = ["compatible-mypy"], version = "^6.0.1"}
types-psycopg2 = "^2.9.21.20260223"
types-nanoid = "^2.0.0.20240601"
celery-types = "^0.26.0"
pandas-stubs = "^3.0.0.260204"
types-requests = "^2.33.0.20260327"
pyproj = "^3.7.2"
types-shapely = "^2.1.0.20250917"
factory_boy = "^3.3.3"
coverage = "^7.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests", "notebooks",
  "batid/services/imports", # Should be added back, but attack vector less likely
  "batid/services/data_gouv_publication.py",
  "batid/services/data_fix"
]
skips = [
  "B113", # Lots of requests without timeout to be fixed, but we need to address more important things first
  "B202" # We do have unsanitized extracts, but from hardcoded sources. It must be fixed, but the attack vector is more intricated
]

[tool.mypy]
exclude = [
  "batid/tests/",
  "api_alpha/tests/",
  "xp/",
  "scripts/",
  "notebooks/",
  "webhook/",
  "manage.py",
]
disallow_untyped_defs = false
plugins = ["mypy_django_plugin.main"]
mypy_path = "stubs"

[[tool.mypy.overrides]]
module = ["boto3.*", "rest_framework", "rest_framework.*", "rest_framework_tracking.*", "rest_framework_gis.*", "geopandas", "taggit.*", "fiona", "authlib.*"]
follow_untyped_imports = true

[tool.django-stubs]
django_settings_module = "app.settings"
